Added new camera/ drawmap system.
new camera system makes the camera just move along whilst the player can
move around the screen. Also implemented a death if player moves off the
left side of the screen. Designed a logo we might want to use and drew
it into the start of the level in the background. Fixed some animation
issues aswell.